requirements.txt
python-dotenv==1.0.0
openai==1.12.0
google-generativeai==0.3.2
pydantic==2.5.2
pandas==2.1.3
openpyxl==3.1.2
python-ffmpeg==2.0.4
moviepy==1.0.3 